requirements.txt
datasets>=3.6.0
transformers>=4.53.0
torch>=2.7.1
numpy>=2.3.1
pandas>=2.3.0
matplotlib>=3.10.3
seaborn>=0.13.2
tqdm>=4.67.1
scikit-learn>=1.5.0
lm-eval>=0.4.9
openai>=1.0.0
anthropic>=0.7.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.0.0
rich>=13.0.0 